+++ frontend/src/utils/GeneralUtils.ts
@@ -10,3 +10,13 @@ export const getFirstNumber = (word: string): string => {
   const numbers = word.match(/^\d+/);
   return numbers ? numbers[0] : "";
 };
+
+export const getReadableDate = (): string => {
+  const date = new Date();
+  const options: Intl.DateTimeFormatOptions = {
+    month: "long",
+    day: "numeric",
+    year: "numeric",
+  };
+  return date.toLocaleDateString("en-US", options);
+};
